Fix problematic line in stats lesson
For some reason, in the section discussing the dim command, the
rendered HTML (and maybe pdf?) included an extra line break in
an unordered list. See the line with (`r nrow(iris)`). Moving
the parenthetical call to nrow to the next line will result in
remaining text being rendered outside of the unordered list.
Don't know why that happens, but this fixes the problem.